Disallow dangerous qrexec operations
Whonix-Workstation should never need to open a file, URL, or
application in anything other than a Whonix-Workstation disposable VM.
Allowing it to even ask to do one of these actions in an unsafe way is
dangerous, since a user who isn't paying enough attention might allow
the request and leak their IP address in so doing (for instance, by
visiting an attacker-provided URL in a VM that has clearnet access).
Don't allow Whonix-Workstation to take (or ask to take) any of these
dangerous actions.