Set NetVM and DispVM properly for StandaloneVMs
Previously, when creating a Whonix-Workstation standalone, the NetVM
would be sys-firewall by default, and the default DispVM would be
default-dvm. Both of these are wrong for most Whonix-Workstation qubes,
so now when a qube has the whonix-ws feature set on it, it will
automtically have its NetVM and DispVM changed to appropriate values for
a Whonix-Workstation machine. This makes standalone creation work as
expected.