Increase default initial memory for Whonix workstation
It appears that Whonix workstation starts more things on boot, and with
default 400M initial memory it's swapping quite a bit, resulting in slow
startup. Increase default memory for it, but only if user hasn't
modified it before already. The template's value is used as default for
app qubes based on it, so it's enough to adjust on the template.

Note this will also apply to existing templates, at next update, as
features are re-requested on each update.