Respect initial property if it references Whonix
I noticed during preload tests that my Whonix qubes properties were
being overridden. Setting a feature on the template is not the best
option, seen that it doesn't consider multi-VM setups where there can be
multiple disposable templates. The order is as follows:

- None if current property is None;
- Current property if it asks Netvm it is a Gateway, itself it it asks
  for a disposable template and it is a Workstation or current property
  if it asks for a disposable template and it is a Workstation;
- Check for features in template chain; and
- Fallback to hardcoded values

Whonix Templates are a bit different, they are created with the global
default_dispvm, which is not Whonix by default, they are also created
before a disposable template (at least a descendant) can exist.
Workstation as the default disposable template, it will be kept,

For: https://github.com/QubesOS/qubes-issues/issues/1512