Replace default qrexec policy on update
It avoids issues when updated policy is left in the .rpmnew file and
user gets a ton of notifications about denied calls. If there were any
user changes, they will be saved in the .rpmsave file.

The file now has a comment discouraging direct edits there, so there is
known method how to affect whonix rules without changing the default
file.
Technically, that comment wasn't there before, so some user might be not
aware of this recommendation (and thus their customizations will get
reverted by the update). But it's rather unlikely event and the fallout
of outdated rules is bad too.